backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]>=1.4
asyncpg==0.29.0
psycopg[binary]==3.2.1
alembic==1.13.3
pgvector==0.3.2

# Redis / queues
redis==5.1.1
bullmq>=2.0.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography==43.0.1

# AI / LangChain ecosystem - ALIGNED VERSIONS
langchain==0.3.7
langchain-core>=0.3.15,<0.4.0
langchain-community==0.3.7        # Must match langchain version
langchain-google-genai==2.0.4
langgraph==0.2.38

# HTTP client (WhatsApp API)
httpx==0.27.2

# Validation
pydantic==2.9.2
pydantic-settings>=2.5.2
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
structlog==24.4.0